+++ src/components/Tabs/index.ts
@@ -1,10 +1,8 @@
-export { Tabs, TAG_CONFIG } from "./Tabs";
+export { Tabs } from "./Tabs";
 export type {
   TabsProps,
   TabItem,
   TabGroupMeta,
   TabOverflowMode,
-  TabTagType,
-  TabTagLabels,
   TabVariant,
 } from "./Tabs";